Creating Crud Slice and Async Thunks

+++ frontend/src/components/CrudComponent.jsx
@@ -1,4 +1,4 @@
-import React, { forwardRef, useEffect, useImperativeHandle, useRef, useState } from 'react';
+import React, { createContext, forwardRef, useCallback, useContext, useEffect, useImperativeHandle, useMemo, useRef } from 'react';
 import '../styles/style.css'
 import { useForm } from 'react-hook-form';
 import { useDispatch, useSelector } from 'react-redux';
@@ -6,10 +6,45 @@ import { Link } from 'react-router-dom';
 import { handleLogout } from '../pages/login/Login';
 import { setFormErrors, setFormValues } from '../helpers/reactHookForm';
 
-const FormComponent = forwardRef(({ name, formFields, createRecord, updateRecord, resetForm }, ref) => {
+
+const ApiContext = createContext(null)
+const useApi = () => useContext(ApiContext)
+
+const FormField = ({ name, register, configs, errors }) => {
+    console.log('Render - FormField')
+    const required = configs.rules.required ? 'required' : ''
+    const errorMsg = errors[name]?.message
+
+    let formField;
+    if (configs.type === 'select') {
+        formField = (
+            <>
+                <select {...register(name, configs.rules)} required={required} >
+                    <option value="" disabled>{`Select ${configs.name}`}</option>
+                    {configs.options.map((opt, i) => <option key={i} value={opt.value}>{opt.label}</option>)}
+                </select>
+            </>
+        )
+
+    } else {
+        formField = (
+            <input {...register(name, configs.rules)} placeholder={`Enter ${configs.name}`} required={required} />
+        )
+    }
+    return (
+        <React.Fragment>
+            <p>{formField}</p>
+            {errorMsg && <p style={{ color: 'red' }}> {errorMsg} </p>}
+        </ React.Fragment>
+    )
+}
+
+const FormComponent = forwardRef(({ formFields }, ref) => {
     console.log('Render - FormComponent');
 
+    const { name, createRecord, updateRecord, resetForm } = useApi()
     const mode = useSelector((state) => state[name].mode)
+    const { __add, __change } = useSelector((state) => state[name].permissions)
     const reactHookForm = useForm({ mode: 'all', defaultValues: formFields.defaultValues })
     const { register, handleSubmit, reset, formState: { errors } } = reactHookForm
 
@@ -17,13 +52,13 @@ const FormComponent = forwardRef(({ name, formFields, createRecord, updateRecord
 
     const dispatch = useDispatch()
     const handleFormSubmit = (data) => {
-        if (mode === 'Update' && data.rec_id) {
+        if (mode === 'Update' && data.rec_id && __change) {
             dispatch(updateRecord({
                 ...data,
                 successCallBack: (response) => { reset() },
                 errorCallBack: (err) => setFormErrors({ reactHookForm, err })
             }))
-        } else if (mode === 'Create') {
+        } else if (mode === 'Create' && __add) {
             dispatch(createRecord({
                 ...data,
                 successCallBack: (response) => { reset() },
@@ -38,73 +73,65 @@ const FormComponent = forwardRef(({ name, formFields, createRecord, updateRecord
     }
     return (
         <form onSubmit={handleSubmit(handleFormSubmit)} noValidate>
-            {Object.entries(formFields.fields).map(([name, configs], index) => {
-                const required = configs.rules.required ? 'required' : ''
-                const errorMsg = errors[name]?.message
-                return (
-                    <React.Fragment key={index}>
-                        <p>
-                            {configs.type === 'select' ? (
-                                <select {...register(name, configs.rules)} required={required} >
-                                    <option value="" disabled>{`Select ${configs.name}`}</option>
-                                    {configs.options.map((opt, i) => <option key={i} value={opt.value}>{opt.label}</option>)}
-                                </select>
-                            ) : (
-                                <input {...register(name, configs.rules)} placeholder={`Enter ${configs.name}`} required={required} />
-                            )}
-                        </p>
-                        {errorMsg && <p style={{ color: 'red' }}> {errorMsg} </p>}
-                    </ React.Fragment>
-                )
-            })}
+            {Object.entries(formFields.fields).map(([name, configs], index) => <FormField key={index} {...{ register, name, configs, errors }} />)}
             <button>{mode === 'Create' ? 'Submit' : 'Update'}</button>
             <button type='button' onClick={handleReset} style={{ marginLeft: 10 }}>Cancel</button>
         </form >
     )
 })
 
-const TableComponent = ({ name, tableFields, reactHookFormRef, fetchSingleRecord, deleteRecord }) => {
-    console.log('Render - TableComponent');
-    const data = useSelector((state) => state[name].data)
+const TableActions = ({ recId, reactHookFormRef }) => {
+    console.log('Render - TableActions')
+
+    const { name, fetchSingleRecord, deleteRecord } = useApi()
+    const { __change, __delete } = useSelector((state) => state[name].permissions)
 
     const dispatch = useDispatch()
-
     const handleEdit = (rec_id) => {
-        const { reactHookForm } = reactHookFormRef.current
-        dispatch(fetchSingleRecord({
-            rec_id,
-            successCallBack: (response) => setFormValues({ reactHookForm, response })
-        }))
+        if (__change) {
+            const { reactHookForm } = reactHookFormRef.current
+            dispatch(fetchSingleRecord({
+                rec_id,
+                successCallBack: (response) => setFormValues({ reactHookForm, response })
+            }))
+        }
     }
 
     const handleDelete = (rec_id) => {
-        const isDelete = confirm('Are you Sure?')
-        if (isDelete) dispatch(deleteRecord({ rec_id }))
+        if (__delete) {
+            const isDelete = confirm('Are you Sure?')
+            if (isDelete) dispatch(deleteRecord({ rec_id }))
+        }
     }
+
+    return <>
+        {__change && <td><button onClick={() => handleEdit(recId)}>edit</button></td>}
+        {__delete && <td><button onClick={() => handleDelete(recId)}>delete</button></td>}
+    </>
+}
+
+const TableComponent = ({ tableFields, reactHookFormRef }) => {
+    console.log('Render - TableComponent');
+    const { name } = useApi()
+    const data = useSelector((state) => state[name].data)
+    const { __change, __delete } = useSelector((state) => state[name].permissions)
+
+    const showActions = __change || __delete
 
     return (
         <table>
             <thead>
                 <tr>
-                    {Object.values(tableFields).map((label, headerIndex) => {
-                        return (
-                            <th key={headerIndex}>{label}</th>
-                        )
-                    })}
-                    <th colSpan={2}>Actions</th>
+                    {Object.values(tableFields).map((label, headerIndex) => <th key={headerIndex}>{label}</th>)}
+                    {showActions && <th colSpan={2}>Actions</th>}
                 </tr>
             </thead>
             <tbody>
                 {data.map((obj, rowIndex) => {
                     return (
                         <tr key={rowIndex}>
-                            {Object.keys(tableFields).map((key, dataIndex) => {
-                                return (
-                                    <td key={dataIndex}>{obj[key]}</td>
-                                )
-                            })}
-                            <td><button onClick={() => handleEdit(obj.rec_id)}>edit</button></td>
-                            <td><button onClick={() => handleDelete(obj.rec_id)}>delete</button></td>
+                            {Object.keys(tableFields).map((key, dataIndex) => <td key={dataIndex}>{obj[key]}</td>)}
+                            {showActions && <TableActions {...{ recId: obj.rec_id, reactHookFormRef }} />}
                         </tr>
                     )
                 })}
@@ -120,7 +147,18 @@ const TitleComponent = ({ title }) => {
     )
 }
 
-export function CrudComponent({ name, fetchData, fetchSingleRecord, createRecord, updateRecord, deleteRecord, resetForm }) {
+const Navigation = () => {
+    return <>
+        <div>
+            <Link to="/dashboard">Dashboard</Link>
+        </div>
+        <div>
+            <button onClick={handleLogout}>Log out</button>
+        </div>
+    </>
+}
+
+export function CrudComponent({ name, api }) {
     console.log('Render - CrudComponent');
 
     const reactHookFormRef = useRef({})
@@ -128,25 +166,30 @@ export function CrudComponent({ name, fetchData, fetchSingleRecord, createRecord
     const title = useSelector((state) => state[name].title)
     const formFields = useSelector((state) => state[name].formFields)
     const tableFields = useSelector((state) => state[name].tableFields)
+    const { __view, __add, __change } = useSelector((state) => state[name].permissions)
+
+    const fetchSingleRecord = useCallback(api.fetchSingleRecord, [])
+    const createRecord = useCallback(api.createRecord, [])
+    const updateRecord = useCallback(api.updateRecord, [])
+    const deleteRecord = useCallback(api.deleteRecord, [])
+    const resetForm = useCallback(api.resetForm, [])
+    const apiValue = useMemo(() => ({ name, fetchSingleRecord, createRecord, updateRecord, deleteRecord, resetForm }), [])
 
     const dispatch = useDispatch()
     useEffect(() => {
-        dispatch(fetchData())
+        dispatch(api.fetchData())
     }, [dispatch])
 
+    const showForm = formFields && (__add || __change) ? true : false
+    const showTable = tableFields && (__view) ? true : false
     return (
-        <>
-            <div>
-                <Link to="/dashboard">Dashboard</Link>
-            </div>
-            <div>
-                <button onClick={handleLogout}>Log out</button>
-            </div>
+        <ApiContext.Provider value={apiValue}>
+            <Navigation />
             {title && <TitleComponent {...{ title }} />}
-            {formFields && <FormComponent ref={reactHookFormRef} {...{ name, formFields, createRecord, updateRecord, resetForm }} />}
+            {showForm && <FormComponent ref={reactHookFormRef} {...{ formFields }} />}
             <hr />
-            {tableFields && <TableComponent {...{ name, tableFields, reactHookFormRef, fetchSingleRecord, deleteRecord }} />}
-        </>
+            {showTable && <TableComponent {...{ tableFields, reactHookFormRef }} />}
+        </ApiContext.Provider>
     )
 }
